+++ src/libs/math.ts
@@ -1,0 +1,39 @@
+/**
+ * Function that returns the sum of two numbers
+ *
+ * @param x The number to add to `y`
+ * @param y The number to add to `x`
+ */
+export const add = (x: number, y: number): number => x + y
+
+/**
+ * Function that returns the difference between two numbers
+ *
+ * @param x The number to use with `y`
+ * @param y The number to use with `x`
+ */
+export const subtract = (x: number, y: number): number => x - y
+
+/**
+ * Function that returns the quotient of two numbers
+ *
+ * @param x The number to divide with `y`
+ * @param y The number to divide with `x`
+ */
+export const divide = (x: number, y: number): number => x / y
+
+/**
+ * Function that returns the product of two numbers
+ *
+ * @param x The number to multiply with `y`
+ * @param y The number to multiply with `x`
+ */
+export const multiply = (x: number, y: number): number => x * y
+
+/**
+ * Function that returns the exponential value of two numbers
+ *
+ * @param x The number to use to `y`
+ * @param y The number to use to `x`
+ */
+export const exp = (x: number, y: number): number => Math.pow(x, y)
