--- src/math.ts
@@ -1,14 +1,0 @@
-/** Function that returns the sum of two numbers */
-export const add = (x: number, y: number): number => x + y
-
-/** Function that returns the difference between two numbers */
-export const subtract = (x: number, y: number): number => x - y
-
-/** Function that returns the quotient of two numbers */
-export const divide = (x: number, y: number): number => x / y
-
-/** Function that returns the product of two numbers */
-export const multiply = (x: number, y: number): number => x * y
-
-/** Function that returns the exponential value of two numbers */
-export const exp = (x: number, y: number): number => Math.pow(x, y)
